modified main classes
